predict_test.py
import tempfile
from pathlib import Path
from typing import Optional
import json

import torch
from PIL import Image
from cog import BasePredictor, Input, Path as CogPath


class Predictor(BasePredictor):
    """Test predictor to verify Cog setup works."""
    
    def setup(self) -> None:
        """Initialize the predictor."""
        # Verify CUDA availability
        if torch.cuda.is_available():
            print(f"CUDA available with {torch.cuda.device_count()} GPU(s)")
            print(f"Using GPU: {torch.cuda.get_device_name()}")
        else:
            print("CUDA not available, using CPU")
        
        print("Test predictor setup complete")
    
    def predict(
        self,
        image: CogPath = Input(description="Input image"),
        confidence_threshold: float = Input(
            description="Confidence threshold", 
            default=0.1, 
            ge=0.0, 
            le=1.0
        )
    ) -> dict:
        """Run a test prediction."""
        
        try:
            # Just verify we can load the image
            with Image.open(image) as img:
                width, height = img.size
                mode = img.mode
            
            return {
                "success": True,
                "message": "Test prediction successful",
                "image_info": {
                    "width": width,
                    "height": height,
                    "mode": mode
                },
                "confidence_threshold": confidence_threshold,
                "mock_prediction": {
                    "species": "Test Species",
                    "confidence": 0.85
                }
            }
            
        except Exception as e:
            return {
                "success": False, 
                "error": f"Test prediction failed: {str(e)}"
            }