tests/test_predict.py
import pytest
import tempfile
from pathlib import Path
from PIL import Image
import numpy as np
import json
from unittest.mock import patch, MagicMock
import subprocess

class TestPredictor:
    """Test version of Predictor that doesn't use Cog decorators."""
    
    def setup(self):
        """Initialize the predictor."""
        print("Test predictor setup complete")
    
    def predict(
        self,
        image,
        confidence_threshold=0.1,
        iou_threshold=0.5,
        country_code=None,
        admin1_region=None,
        model_version="v4.0.1a",
        run_mode="full_ensemble",
        return_top_k=5
    ):
        """Test prediction method."""
        # Create temporary directory for processing
        with tempfile.TemporaryDirectory() as temp_dir:
            temp_path = Path(temp_dir)
            
            # Copy input image to temp directory
            input_path = temp_path / "input.jpg"
            
            # Handle different input types
            if isinstance(image, (str, Path)):
                # Copy file
                import shutil
                shutil.copy2(image, input_path)
            else:
                # Assume it's a file-like object
                with open(input_path, 'wb') as f:
                    f.write(image.read())
            
            # Verify image can be opened
            try:
                with Image.open(input_path) as img:
                    # Convert to RGB if needed
                    if img.mode != 'RGB':
                        img = img.convert('RGB')
                        img.save(input_path)
            except Exception as e:
                return {"error": f"Invalid image file: {e}"}
            
            # For testing, return a mock successful response
            return {
                "success": True,
                "model_version": model_version,
                "prediction_source": "test_mock",
                "final_prediction": {
                    "species": "Test Species",
                    "confidence": 0.85
                },
                "detections": [
                    {
                        "category": "animal",
                        "confidence": 0.9,
                        "bbox": [0.1, 0.2, 0.3, 0.4]
                    }
                ],
                "classifications": [
                    {"species": "Test Species", "confidence": 0.85},
                    {"species": "Alternative Species", "confidence": 0.15}
                ]
            }


@pytest.fixture
def predictor():
    """Create a test predictor instance."""
    pred = TestPredictor()
    pred.setup()
    return pred


@pytest.fixture
def test_image():
    """Create a test image."""
    # Create a simple test image
    image = Image.new('RGB', (640, 480), color='green')
    
    with tempfile.NamedTemporaryFile(suffix='.jpg', delete=False) as f:
        image.save(f.name)
        yield f.name
    
    # Cleanup
    Path(f.name).unlink()


def test_predict_basic(predictor, test_image):
    """Test basic prediction functionality."""
    result = predictor.predict(image=test_image)
    
    assert "success" in result
    assert result["success"] is True
    assert "model_version" in result
    assert "final_prediction" in result
    assert result["final_prediction"]["species"] == "Test Species"


def test_predict_with_country_code(predictor, test_image):
    """Test prediction with country code."""
    result = predictor.predict(
        image=test_image,
        country_code="USA",
        confidence_threshold=0.1
    )
    
    assert "success" in result
    assert result["success"] is True
    assert "final_prediction" in result


def test_predict_detector_only(predictor, test_image):
    """Test detector-only mode."""
    result = predictor.predict(
        image=test_image,
        run_mode="detector_only"
    )
    
    assert "success" in result
    assert result["success"] is True
    # Should have detections
    assert "detections" in result
    assert len(result["detections"]) > 0


def test_invalid_image(predictor):
    """Test handling of invalid image."""
    # Create invalid image file
    with tempfile.NamedTemporaryFile(suffix='.txt', delete=False) as f:
        f.write(b"not an image")
        
        result = predictor.predict(image=f.name)
        assert "error" in result
        
    Path(f.name).unlink()


if __name__ == "__main__":
    pytest.main([__file__])