requirements.txt
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
numpy==1.24.3
speciesnet==5.0.0